add owner

+++ controllers/contactsController.js
@@ -1,22 +1,31 @@
 const { Contact } = require("../models/contactsModel");
-const {createNotFoundError} = require("../helpers")
+const { createNotFoundError } = require("../helpers");
 
 const getAllContacts = async (req, res, next) => {
-  const contacts = await Contact.find({});
+  const { _id } = req.user;
+  const contacts = await Contact.find({ owner: _id });
   return res.status(200).json({
     message: contacts,
   });
 };
 
 const addContact = async (req, res, next) => {
-  const createdContact = await Contact.create(req.body);
+  const { name, email, phone } = req.body;
+  const { _id } = req.user;
+  const createdContact = await Contact.create({
+    name,
+    email,
+    phone,
+    owner: _id,
+  });
   return res.status(201).json({ message: createdContact });
 };
 
 const removeContactById = async (req, res, next) => {
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndDelete(contactId);
-  if (result) {  
+  const { _id } = req.user;
+  const result = await Contact.findByIdAndDelete(contactId, { owner: _id });
+  if (result) {
     return res.status(200).json({ message: `${result.name} deleted` });
   }
   return next(createNotFoundError());
@@ -24,7 +33,8 @@ const removeContactById = async (req, res, next) => {
 
 const getContactByID = async (req, res, next) => {
   const { contactId } = req.params;
-  const contact = await Contact.findById(contactId);
+  const { _id } = req.user;
+  const contact = await Contact.findById(contactId, { owner: _id });
   if (contact) {
     return res.status(200).json({ message: contact });
   }
@@ -33,7 +43,10 @@ const getContactByID = async (req, res, next) => {
 
 const updateContact = async (req, res, next) => {
   const { contactId } = req.params;
-  const result = await Contact.findByIdAndUpdate(contactId, req.body);
+  const { _id } = req.user;
+  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
+    owner: _id,
+  });
   if (result) {
     return res.status(200).json({ message: `${result.name} was updated` });
   }
@@ -43,9 +56,10 @@ const updateContact = async (req, res, next) => {
 const updateContactStatus = async (req, res, next) => {
   const { contactId } = req.params;
   const { favorite } = req.body;
+  const { _id } = req.user;
   const updatedContact = await Contact.findByIdAndUpdate(
     contactId,
-    { favorite },
+    { owner: _id, favorite },
     { new: true }
   );
   if (updatedContact) {
